+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1.6.2 (not released yet)
 - Support Team DM scores above 255
 - Added additional textures from base game to level editor texture browser (contribution from Goober)
 - Display proper server name in the scoreboard when joining by link
+- Redesign level auto-downloader UI and add rejoining the server after download has completed
 
 Version 1.6.1 (released 2021-02-20)
 -----------------------------------